refactor: moving input_data to io module since onyl helper struct holding input data
