+++ src/types/native/record.rs
@@ -1,8 +1,4 @@
-use super::contour_point::detect_delimiter;
-use csv::ReaderBuilder;
 use serde::Deserialize;
-use std::fs::File;
-use std::path::Path;
 
 #[derive(Debug, Clone, PartialEq, Deserialize)]
 pub struct Record {
@@ -12,20 +8,4 @@ pub struct Record {
     pub measurement_1: Option<f64>,
     #[serde(deserialize_with = "csv::invalid_option")]
     pub measurement_2: Option<f64>,
-}
-
-pub fn read_records<P: AsRef<Path>>(path: P) -> anyhow::Result<Vec<Record>> {
-    let delim = detect_delimiter(&path)?;
-    let file = File::open(path)?;
-    let mut reader = ReaderBuilder::new()
-        .delimiter(delim)
-        .has_headers(true)
-        .from_reader(file);
-
-    let mut records = Vec::new();
-    for result in reader.deserialize() {
-        let record: Record = result?;
-        records.push(record);
-    }
-    Ok(records)
 }
